Fix incorrect cl-case, to silence warnings
This commit does not modify code behavior at all but it does silence
the following warning on initialization, which says the following for
the three cases of symbol, word, and regexp:

  Case 'symbol will match ‘quote’.  If that’s intended, write (symbol
quote) instead. Otherwise, don’t quote ‘symbol’.

That warning is pointing out that this there is no need to quote the
symbols presented in the KEYLIST part of the CL-CASE form. Quoting
them has the side-effect that CL-CASE would incorrectly match the
symbol quote.

The behavior can be seen below:

ELISP> (defvar *a/t1* 'symbol)
*a/t1*
ELISP> (defvar *a/t2* 'quote)
*a/t2*

;; desired behavior, after this commit
ELISP> (cl-case *a/t1* (word "word") (symbol "symbol") (t "default"))
"symbol"
ELISP> (cl-case *a/t2* (word "word") (symbol "symbol") (t "default"))
"default"

;; undesired behavior, before this commit
ELISP> (cl-case *a/t1* ('word "word") ('symbol "symbol") (t "default"))
"symbol"
ELISP> (cl-case *a/t2* ('word "word") ('symbol "symbol") (t "default"))
"word". ;; this is undesired! we want this to be default

In practice this is not a problem now because the variable
iedit-occurrence-type-local is never set to the symbol quote.

So the only negative effect of the current code is to generate the
warning. And the only effect of this change is to suppress the
warning.